define new routes and optimized it and make it reuasble components
